Move authentication to new utils dir
